Switch to vfloat4 and remove float4
The main purpose of this change is to have a cleaner mapping
on to SIMD instructions, as the compilers only seem to do a
good job when directly operating on __m128 types.

This stage in the implementation removes all float4 code,
and users relatively direct replacements without other
algorithmic improvements. The resulting binary code is ~10KB
smaller for broadly similar peformance (perhaps slightly
slower, due to use of lane inserts).